Fix requirements.txt syntax
Existing version of the `requirements.txt` file could not be used by `pip`.

Note that I'm not using `--no-dependencies` option for installing `tensorvision` and `tensorboard` as it would require putting `--install-option="--no-deps"` which leads to wheels being disabled for the entire file (some packets here could not be installed without wheels support). Not sure if this is critical. If so, the installation procedure has to be split into multiple stages/requirements lists.